+++ AmetrinTest/src/main/java/com/ametrinstudios/ametrin_test/data/provider/TestBlockTagsProvider.java
@@ -3,6 +3,7 @@ package com.ametrinstudios.ametrin_test.data.provider;
 import com.ametrinstudios.ametrin.data.provider.ExtendedBlockTagsProvider;
 import com.ametrinstudios.ametrin_test.AmetrinTestMod;
 import com.ametrinstudios.ametrin_test.registry.TestBlocks;
+import com.ametrinstudios.ametrin_test.registry.TestTags;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.tags.BlockTags;
@@ -23,6 +24,10 @@ public final class TestBlockTagsProvider extends ExtendedBlockTagsProvider {
 
         tag(BlockTags.LOGS_THAT_BURN).add(
                 TestBlocks.TEST_LOG.get()
+        );
+
+        tag(TestTags.Blocks.TEST_PORTAL_FRAMES).add(
+                TestBlocks.TEST_BLOCK.get()
         );
     }
 }