+++ AmetrinTest/src/main/java/com/ametrinstudios/ametrin_test/data/provider/loot/TestBlockLootProvider.java
@@ -2,12 +2,13 @@ package com.ametrinstudios.ametrin_test.data.provider.loot;
 
 import com.ametrinstudios.ametrin.data.provider.loot_table.ExtendedBlockLootSubProvider;
 import com.ametrinstudios.ametrin_test.registry.TestBlocks;
+import com.ametrinstudios.ametrin_test.registry.TestItems;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.world.level.block.Block;
 import org.jetbrains.annotations.NotNull;
 
-public final class TestBlockLootSubProvider extends ExtendedBlockLootSubProvider {
-    public TestBlockLootSubProvider(HolderLookup.Provider registries) {
+public final class TestBlockLootProvider extends ExtendedBlockLootSubProvider {
+    public TestBlockLootProvider(HolderLookup.Provider registries) {
         super(registries);
     }
 
@@ -17,6 +18,9 @@ public final class TestBlockLootSubProvider extends ExtendedBlockLootSubProvider
                 TestBlocks.TEST_BLOCK,
                 TestBlocks.TEST_LOG
         );
+
+        dropOther(TestBlocks.TEST_SKULL, TestItems.TEST_SKULL);
+        dropOther(TestBlocks.TEST_SKULL_WALL, TestItems.TEST_SKULL);
     }
 
     @Override